refactor: ♻️ explode api clients into main application
