+++ apps/web/package.json
@@ -9,6 +9,7 @@
     "dev": "dotenv-run -- vite dev --clearScreen false",
     "e2e": "dotenv-run -- playwright test",
     "e2e:ui": "dotenv-run -- playwright test --ui",
+    "gen": "openapi-typescript --root-types",
     "lint": "eslint --cache .",
     "preview": "node .output/server/index.mjs",
     "test": "vitest",
@@ -16,17 +17,18 @@
   },
   "dependencies": {
     "@hookform/resolvers": "5.1.1",
-    "@popcorn.fyi/api-clients": "workspace:*",
     "@tanstack/match-sorter-utils": "8.19.4",
     "@tanstack/react-query": "5.81.5",
     "@tanstack/react-router": "1.122.0",
     "@tanstack/react-router-with-query": "1.122.0",
     "@tanstack/react-start": "1.122.0",
     "@tanstack/react-table": "8.21.3",
+    "@upstash/redis": "1.35.0",
     "@vercel/analytics": "1.5.0",
     "@vercel/speed-insights": "1.2.0",
     "clsx": "2.1.1",
     "holy-loader": "2.3.13",
+    "openapi-fetch": "0.14.0",
     "react": "19.1.0",
     "react-dom": "19.1.0",
     "react-hook-form": "7.59.0",
@@ -57,6 +59,7 @@
     "babel-plugin-react-compiler": "19.1.0-rc.2",
     "happy-dom": "18.0.1",
     "msw": "2.10.2",
+    "openapi-typescript": "7.8.0",
     "picocolors": "1.1.1",
     "tailwindcss": "4.1.11",
     "vite": "6.3.5",
