+++ apps/web/src/components/people/person-card.tsx
@@ -1,5 +1,6 @@
-import { tmdbImageUrl } from "@popcorn.fyi/api-clients/utils";
 import { Link } from "@tanstack/react-router";
+
+import { tmdbImageUrl } from "@/lib/api-clients/urls";
 
 import { CardImageFallback } from "../media/card-image-fallback";
 import { MediaType } from "../media/media-type";
