+++ apps/web/src/components/tv/tv-show-hero.tsx
@@ -1,5 +1,6 @@
-import { tmdbImageUrl } from "@popcorn.fyi/api-clients/utils";
 import { Link } from "@tanstack/react-router";
+
+import { tmdbImageUrl } from "@/lib/api-clients/urls";
 
 import { MediaType } from "../media/media-type";
 import { TrendingBadge } from "../media/trending-badge";
