+++ apps/web/src/lib/api-clients/wikidata.gen.d.ts
