+++ apps/web/src/lib/tmdb.ts
@@ -1,6 +1,5 @@
-import { setupClient } from "@popcorn.fyi/api-clients/tmdb-v3";
-
 import env from "../../env.config";
+import { setupClient } from "./api-clients/tmdb-v3";
 
 export const client = setupClient(env.TMDB_API_TOKEN, {
   token: env.KV_REST_API_TOKEN,
