+++ apps/web/src/routes/_layout/people.$id.tsx
@@ -1,4 +1,3 @@
-import { tmdbImageUrl } from "@popcorn.fyi/api-clients/utils";
 import { useSuspenseQuery } from "@tanstack/react-query";
 import { createFileRoute } from "@tanstack/react-router";
 import * as v from "valibot";
@@ -7,6 +6,7 @@ import { personDetailsOptions } from "@/api/people/details";
 import { personExternalOptions } from "@/api/people/details.external";
 import { PersonDetails } from "@/components/people/person-details";
 import { site } from "@/config/site";
+import { tmdbImageUrl } from "@/lib/api-clients/urls";
 import { seo } from "@/lib/seo";
 import { PathParamsSchema } from "@/schemas/path-params";
