+++ codecov.yaml
@@ -14,7 +14,6 @@ coverage:
 flags:
   monorepo:
     paths:
-      - libs/api-clients
       - apps/web
 
 comment:
