--- libs/api-clients/.lefthook.json
@@ -1,12 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/lefthook.json",
-  "pre-commit": {
-    "commands": {
-      "lint": {
-        "root": "libs/tmdb",
-        "run": "pnpm eslint --cache --quiet {staged_files}",
-        "stage_fixed": true
-      }
-    }
-  }
-}
