--- libs/api-clients/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "@popcorn.fyi/api-clients",
-  "private": true,
-  "exports": {
-    "./utils": "./src/index.ts",
-    "./tmdb-v3": "./src/tmdb-v3.ts",
-    "./wikidata": "./src/wikidata.ts"
-  },
-  "scripts": {
-    "coverage": "vitest run --coverage",
-    "gen": "npx openapi-typescript --root-types",
-    "lint": "eslint --cache .",
-    "test": "vitest",
-    "typecheck": "tsc"
-  },
-  "dependencies": {
-    "@upstash/redis": "1.35.0",
-    "openapi-fetch": "0.14.0"
-  },
-  "devDependencies": {
-    "@popcorn.fyi/eslint": "workspace:*",
-    "@popcorn.fyi/ts": "workspace:*",
-    "openapi-typescript": "7.8.0"
-  }
-}
