--- libs/api-clients/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "@popcorn.fyi/ts/base.json",
-  "compilerOptions": {
-    "types": ["vitest/globals"]
-  },
-  "include": ["*.config.ts", "src"]
-}
